Idempotent db seed
Eventually we'll introduce some sort of data admin views. But until then
I'd like to store everything in seed. But I don't want to re-load all
the same data all the time.

Provide an idempotent way to add a host to an episode. Remove the
commands to drop all existing data.