Don't wipe tempfiles at end of block
This code was suggested by the Tempfile docs, but it does not work with
ActiveStorage. I _think_ ActiveStorage is multi threaded and the end
result is that the tempfile is not there when Storage is trying to do
its thing. Code works fine without this cleanup and I don't see that the
tempfiles are hanging around.